+++ massiv-test/src/Test/Massiv/Array/Delayed.hs
@@ -32,14 +32,13 @@ import Data.List as L
 
 -- | Alternative implementation of `stackSlicesM` with `concat'`. Useful for testing and benchmarks
 stackSlices' ::
-     (Functor f, Foldable f, Resize r, Source r e, Index ix, Load r (Lower ix) e)
+     (Functor f, Foldable f, Source r e, Index ix, Load r (Lower ix) e)
   => Dim
   -> f (Array r (Lower ix) e)
   -> Array DL ix e
 stackSlices' dim arrsF =
   let fixupSize arr = resize' (Sz (insertDim' (unSz (size arr)) dim 1)) arr
    in concat' dim $ fmap fixupSize arrsF
-
 
 compareAsListAndLoaded ::
      (Eq e, Show e, Foldable (Array r' Ix1), Load r' Ix1 e) => Array r' Ix1 e -> [e] -> Property